refactor(gateway): extract GatewayRuntimeConfigMixin from run.py (slice 20 of #54962)
